[bazel] Sync with latest changes on main.

+++ include/parlay/internal/scheduler_plugins/cilkplus.h
@@ -22,10 +22,10 @@ inline void par_do(Lf left, Rf right, bool) {
 template <typename F>
 inline void parallel_for(size_t start, size_t end, F f,
                          long granularity,
-			 bool) {
+			                   bool) {
   if (granularity == 0)
     cilk_for(size_t i=start; i<end; i++) f(i);
-  else if ((end - start) <= granularity)
+  else if ((end - start) <= static_cast<size_t>(granularity))
     for (size_t i=start; i < end; i++) f(i);
   else {
     size_t n = end-start;
@@ -40,4 +40,3 @@ inline void parallel_for(size_t start, size_t end, F f,
 
 #endif
 #endif  // PARLAY_INTERNAL_SCHEDULER_PLUGINS_CILK_HPP_
-
